+++ lib/kitchen/directions/move_solutions_to_answer_key/move_solutions_from_exercise_section.rb
@@ -8,13 +8,7 @@ module Kitchen::Directions::MoveSolutionsFromExerciseSection
 
   class V1
     def bake(chapter:, append_to:, section_class:, title_number:)
-      solutions_clipboard = Kitchen::Clipboard.new
-      chapter.search("section.#{section_class}").exercises.each do |exercise|
-        solution = exercise.solution
-        next unless solution
-
-        solution.cut(to: solutions_clipboard)
-      end
+      solutions_clipboard = chapter.search("section.#{section_class}").solutions.cut
 
       return if solutions_clipboard.items.empty?