+++ lib/kitchen/selectors/base.rb
@@ -56,6 +56,12 @@ module Kitchen
       # Selector for an unit
       # @return [String]
       attr_accessor :unit
+      # Selector for an unit
+      # @return [String]
+      attr_accessor :solution
+      # Selector for an injected question
+      # @return [String]
+      attr_accessor :injected_question
 
       # Override specific selectors
       #