+++ lib/kitchen/selectors/standard_1.rb
@@ -26,6 +26,9 @@ module Kitchen
         self.example                    = "div[data-type='example']"
         self.exercise                   = "div[data-type='exercise']"
         self.unit                       = "div[data-type='unit']"
+        self.solution                   = "div[data-type='solution'], " \
+                                          "div[data-type='question-solution']"
+        self.injected_question          = "div[data-type='exercise-question']"
       end
 
     end