+++ spec/directions/bake_notes/bake_numbered_notes/v2_spec.rb
@@ -65,6 +65,17 @@ RSpec.describe Kitchen::Directions::BakeNumberedNotes::V2 do
                 </div>
               </div>
             </div>
+            <div data-type="note" id="123" class="foo">
+              <p>note with injected exercise</p>
+              <div data-type="injected-exercise">
+                <div data-type="exercise-question">
+                  <div data-type="question-stem">a question stem</div>
+                  <div data-type="question-solution">
+                    some solution
+                  </div>
+                </div>
+              </div>
+            </div>
           </div>
         </div>
       HTML
@@ -205,6 +216,29 @@ RSpec.describe Kitchen::Directions::BakeNumberedNotes::V2 do
                       <span class="os-divider">. </span>
                       <div class="os-solution-container">
                         <p>second solution content</p>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div class="foo" data-type="note" id="123">
+                <h3 class="os-title">
+                  <span class="os-title-label">Bar</span>
+                  <span class="os-number">#4</span>
+                </h3>
+                <div class="os-note-body">
+                  <p>note with injected exercise</p>
+                  <div data-type="injected-exercise">
+                    <div class="unnumbered os-hasSolution" data-type="exercise-question">
+                      <div class="os-problem-container">
+                        <div data-type="question-stem">a question stem</div>
+                      </div>
+                      <div data-type="question-solution">
+                        <a class="os-number" href="#exercise-ref">4</a>
+                        <span class="os-divider">. </span>
+                        <div class="os-solution-container">
+                    some solution
+                  </div>
                       </div>
                     </div>
                   </div>